+++ inc/serializeproto/ParserProto.h
@@ -85,7 +85,7 @@ private:
     void skip(WireType wireType);
 
     const char*         m_ptr = nullptr;
-    int                 m_size = 0;
+    ssize_t             m_size = 0;
     IParserVisitor&     m_visitor;
 
     std::uint32_t       m_tag = 0;
