+++ inc/variant/VariantValueStruct.h
@@ -57,7 +57,7 @@ private:
     virtual bool add(const std::string& name, Variant&& variant) override;
     virtual bool add(const Variant& variant) override;
     virtual bool add(Variant&& variant) override;
-    virtual int size() const override;
+    virtual ssize_t size() const override;
     virtual void visit(IVariantVisitor& visitor, Variant& variant, int index, int level, int size, const std::string& name) override;
